Updatd: Weight is now Double

+++ Calorie Hunter/Models/UserProfile.swift
@@ -3,7 +3,7 @@ import Foundation
 struct UserProfile: Codable {
     var name: String
     var dailyCalorieGoal: Int
-    var startWeight: Int
-    var currentWeight: Int
-    var goalWeight: Int
+    var startWeight: Double
+    var currentWeight: Double
+    var goalWeight: Double
 }
